+++ ecoevo/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.1.7.dev3"
+__version__ = "0.1.7.dev4"
 
 from ecoevo.env import EcoEvo